Moving RotationMatrixFactory members in corresponding AxisRotation classes.
