+++ RubbikCube.Core/Model/Rotations/YAxis/YAxisRotation.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 
 namespace RubiksCube.Core.Model.Rotations.YAxis
 {
@@ -15,9 +16,27 @@ namespace RubiksCube.Core.Model.Rotations.YAxis
             return face.GetRow(LayerType);
         }
 
-        public override double[,] GetRotationMatrix(double angle)
+        public override double[,] CreateRotationMatrix(double angle)
         {
-            return RotationMatrixFactory.CreateYRotationMatrix(angle);
+            var matrix = new double[4, 4];
+            matrix[0, 0] = 1;
+            matrix[0, 1] = 0;
+            matrix[0, 2] = 0;
+            matrix[0, 3] = 0;
+            matrix[1, 0] = 0;
+            matrix[1, 1] = Math.Cos(angle);
+            matrix[1, 2] = Math.Sin(angle);
+            matrix[1, 3] = 0;
+            matrix[2, 0] = 0;
+            matrix[2, 1] = -Math.Sin(angle);
+            matrix[2, 2] = Math.Cos(angle);
+            matrix[2, 3] = 0;
+            matrix[3, 0] = 0;
+            matrix[3, 1] = 0;
+            matrix[3, 2] = 0;
+            matrix[3, 3] = 1;
+
+            return matrix;
         }
 
         protected static void FlipPosition(Facie facie)
